+++ covariance/covarianceXsec.h
@@ -37,7 +37,7 @@ class covarianceXsec : public covarianceBase {
     std::string GetParSplineName(const int i) {return _fSplineNames[i];}
 
     //DB Get spline parameters depending on given DetID
-    const std::vector<size_t> GetGlobalSystIndexFromDetID(const int DetID, const SystType Type);
+    const std::vector<int> GetGlobalSystIndexFromDetID(const int DetID, const SystType Type);
     /// @brief EM: value at which we cap spline knot weight
     /// @param i spline parameter index, not confuse with global index
     inline double GetParSplineKnotUpperBound(const int i) {return SplineParams.at(i)._SplineKnotUpBound;}
@@ -53,7 +53,7 @@ class covarianceXsec : public covarianceBase {
     const std::vector<std::string> GetParsNamesFromDetID(const int DetID, const SystType Type);
     /// @brief DB Grab the parameter indices for the relevant DetID
     /// @param Type Type of syst, for example kNorm, kSpline etc
-    const std::vector<size_t> GetParsIndexFromDetID(const int DetID, const SystType Type);
+    const std::vector<int> GetParsIndexFromDetID(const int DetID, const SystType Type);
 
     /// @brief DB Get spline parameters depending on given DetID
     const std::vector<std::string> GetSplineParsNamesFromDetID(const int DetID);
@@ -64,7 +64,7 @@ class covarianceXsec : public covarianceBase {
     const std::vector< std::vector<int> > GetSplineModeVecFromDetID(const int DetID);
     /// @brief Grab the index of the syst relative to global numbering.
     /// @param Type Type of syst, for example kNorm, kSpline etc
-    const std::vector<size_t> GetSystIndexFromDetID(const int DetID, const SystType Type);
+    const std::vector<int> GetSystIndexFromDetID(const int DetID, const SystType Type);
     /// @brief DB Get norm/func parameters depending on given DetID
     const std::vector<XsecNorms4> GetNormParsFromDetID(const int DetID);
 
@@ -80,7 +80,7 @@ class covarianceXsec : public covarianceBase {
     }
     /// @brief Get nominal for a given param
     /// @param i parameter index
-    inline double getNominal(const size_t i) override { return _fPreFitValue.at(i); };
+    inline double getNominal(const int i) override { return _fPreFitValue.at(i); };
 
     /// @brief Checks if parameter belongs to a given group
     /// @param i parameter index
